+++ package.json
@@ -10,13 +10,16 @@
     "@supabase/realtime-js": "^2.10.4",
     "@supabase/ssr": "^0.4.0",
     "@supabase/supabase-js": "^2.44.3",
+    "@types/leaflet": "^1.9.12",
     "autoprefixer": "10.4.17",
     "dotenv": "^16.4.5",
     "geist": "^1.2.1",
+    "leaflet": "^1.9.4",
     "next": "latest",
     "postcss": "8.4.33",
     "react": "18.2.0",
     "react-dom": "18.2.0",
+    "react-leaflet": "^4.2.1",
     "react-nice-avatar": "^1.5.0",
     "socket.io": "^4.7.5",
     "socket.io-client": "^4.7.5",
